Move recipes into their own files in the 'recipes' directory
