+++ src/main/java/io/github/pylonmc/pylon/base/BaseItems.java
@@ -10,9 +10,6 @@ import io.github.pylonmc.pylon.base.content.machines.fluid.FluidPlacer;
 import io.github.pylonmc.pylon.base.content.machines.fluid.FluidVoider;
 import io.github.pylonmc.pylon.base.content.machines.fluid.PortableFluidTank;
 import io.github.pylonmc.pylon.base.content.machines.fluid.WaterPump;
-import io.github.pylonmc.pylon.base.content.machines.simple.Grindstone;
-import io.github.pylonmc.pylon.base.content.machines.simple.MagicAltar;
-import io.github.pylonmc.pylon.base.content.machines.simple.MixingPot;
 import io.github.pylonmc.pylon.base.content.tools.HealthTalisman;
 import io.github.pylonmc.pylon.base.content.machines.hydraulics.HydraulicGrindstoneTurner;
 import io.github.pylonmc.pylon.base.content.machines.hydraulics.HydraulicHammerHead;
@@ -32,6 +29,9 @@ import io.github.pylonmc.pylon.base.content.machines.fluid.Sprinkler;
 import io.github.pylonmc.pylon.base.content.tools.WateringCan;
 import io.github.pylonmc.pylon.base.content.combat.BeheadingSword;
 import io.github.pylonmc.pylon.base.content.combat.RecoilArrow;
+import io.github.pylonmc.pylon.base.recipes.GrindstoneRecipe;
+import io.github.pylonmc.pylon.base.recipes.MagicAltarRecipe;
+import io.github.pylonmc.pylon.base.recipes.MixingPotRecipe;
 import io.github.pylonmc.pylon.base.util.BaseUtils;
 import io.github.pylonmc.pylon.core.config.Settings;
 import io.github.pylonmc.pylon.core.content.fluid.FluidPipe;
@@ -79,7 +79,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, COPPER_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.COPPER_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.COPPER_DUST,
                 new ItemStack(Material.COPPER_INGOT),
                 COPPER_DUST,
@@ -94,7 +94,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, CRUSHED_RAW_COPPER);
         BasePages.RESOURCES.addItem(BaseKeys.CRUSHED_RAW_COPPER);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.CRUSHED_RAW_COPPER,
                 new ItemStack(Material.RAW_COPPER),
                 CRUSHED_RAW_COPPER,
@@ -109,7 +109,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, GOLD_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.GOLD_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.GOLD_DUST,
                 new ItemStack(Material.GOLD_INGOT),
                 GOLD_DUST,
@@ -124,7 +124,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, CRUSHED_RAW_GOLD);
         BasePages.RESOURCES.addItem(BaseKeys.CRUSHED_RAW_GOLD);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.CRUSHED_RAW_GOLD,
                 new ItemStack(Material.RAW_GOLD),
                 CRUSHED_RAW_COPPER,
@@ -139,7 +139,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, IRON_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.IRON_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.IRON_DUST,
                 new ItemStack(Material.IRON_INGOT),
                 IRON_DUST,
@@ -154,7 +154,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, QUARTZ_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.QUARTZ_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.QUARTZ_DUST,
                 new ItemStack(Material.QUARTZ),
                 QUARTZ_DUST,
@@ -169,7 +169,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, DIAMOND_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.DIAMOND_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.DIAMOND_DUST,
                 new ItemStack(Material.DIAMOND),
                 DIAMOND_DUST,
@@ -184,7 +184,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, EMERALD_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.EMERALD_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.EMERALD_DUST,
                 new ItemStack(Material.EMERALD),
                 EMERALD_DUST,
@@ -199,7 +199,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, CRUSHED_RAW_IRON);
         BasePages.RESOURCES.addItem(BaseKeys.CRUSHED_RAW_IRON);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.CRUSHED_RAW_IRON,
                 new ItemStack(Material.RAW_IRON),
                 CRUSHED_RAW_IRON,
@@ -221,7 +221,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, SILVER_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.SILVER_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.SILVER_DUST,
                 SILVER_INGOT,
                 SILVER_DUST,
@@ -250,7 +250,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, ZINC_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.ZINC_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.ZINC_DUST,
                 ZINC_INGOT,
                 ZINC_DUST,
@@ -265,7 +265,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, ROCK_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.ROCK_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.ROCK_DUST,
                 new ItemStack(Material.COBBLESTONE),
                 ROCK_DUST.asQuantity(2),
@@ -294,7 +294,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, LEAD_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.LEAD_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.LEAD_DUST,
                 LEAD_INGOT,
                 LEAD_DUST,
@@ -323,7 +323,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, TIN_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.TIN_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.TIN_DUST,
                 TIN_INGOT,
                 TIN_DUST,
@@ -345,7 +345,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, COBALT_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.COBALT_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.COBALT_DUST,
                 COBALT_INGOT,
                 COBALT_DUST,
@@ -367,7 +367,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, NICKEL_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.NICKEL_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.NICKEL_DUST,
                 NICKEL_INGOT,
                 NICKEL_DUST,
@@ -382,7 +382,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, COAL_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.COAL_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 baseKey("coal_dust_from_coal"),
                 new ItemStack(Material.COAL),
                 COAL_DUST,
@@ -390,7 +390,7 @@ public final class BaseItems {
                 Material.COAL_BLOCK.createBlockData()
         ));
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 baseKey("coal_dust_from_charcoal"),
                 new ItemStack(Material.CHARCOAL),
                 COAL_DUST,
@@ -427,7 +427,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, BRONZE_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.BRONZE_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.BRONZE_DUST,
                 BRONZE_INGOT,
                 BRONZE_DUST,
@@ -449,7 +449,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, BRASS_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.BRASS_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.BRASS_DUST,
                 BRASS_INGOT,
                 BRASS_DUST,
@@ -471,7 +471,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, STEEL_DUST);
         BasePages.RESOURCES.addItem(BaseKeys.STEEL_DUST);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.STEEL_DUST,
                 STEEL_INGOT,
                 STEEL_DUST,
@@ -1171,7 +1171,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, FLOUR);
         BasePages.RESOURCES.addItem(BaseKeys.FLOUR);
 
-        Grindstone.Recipe.RECIPE_TYPE.addRecipe(new Grindstone.Recipe(
+        GrindstoneRecipe.RECIPE_TYPE.addRecipe(new GrindstoneRecipe(
                 BaseKeys.FLOUR,
                 new ItemStack(Material.WHEAT, 2),
                 FLOUR,
@@ -1189,7 +1189,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, DOUGH);
         BasePages.RESOURCES.addItem(BaseKeys.DOUGH);
 
-        MixingPot.Recipe.RECIPE_TYPE.addRecipe(new MixingPot.Recipe(
+        MixingPotRecipe.RECIPE_TYPE.addRecipe(new MixingPotRecipe(
                 BaseKeys.DOUGH,
                 Map.of(new RecipeChoice.ExactChoice(FLOUR), 1),
                 DOUGH,
@@ -1297,7 +1297,7 @@ public final class BaseItems {
         BasePages.RESOURCES.addItem(BaseKeys.COVALENT_BINDER);
         ItemStack output = COVALENT_BINDER.clone();
         output.setAmount(6);
-        MixingPot.Recipe.RECIPE_TYPE.addRecipe(new MixingPot.Recipe(
+        MixingPotRecipe.RECIPE_TYPE.addRecipe(new MixingPotRecipe(
                 BaseKeys.COVALENT_BINDER,
                 Map.of(
                         new RecipeChoice.ExactChoice(new ItemStack(Material.GUNPOWDER)), 4,
@@ -1319,7 +1319,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, SHIMMER_DUST_2);
         BasePages.RESOURCES.addItem(BaseKeys.SHIMMER_DUST_2);
 
-        MixingPot.Recipe.RECIPE_TYPE.addRecipe(new MixingPot.Recipe(
+        MixingPotRecipe.RECIPE_TYPE.addRecipe(new MixingPotRecipe(
                 BaseKeys.SHIMMER_DUST_2,
                 Map.of(
                         new RecipeChoice.ExactChoice(SHIMMER_DUST_1), 1,
@@ -1340,7 +1340,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, SHIMMER_DUST_3);
         BasePages.RESOURCES.addItem(BaseKeys.SHIMMER_DUST_3);
 
-        MagicAltar.Recipe.RECIPE_TYPE.addRecipe(new MagicAltar.Recipe(
+        MagicAltarRecipe.RECIPE_TYPE.addRecipe(new MagicAltarRecipe(
                 BaseKeys.SHIMMER_DUST_3,
                 new ArrayList<>(Arrays.asList(
                         new RecipeChoice.ExactChoice(new ItemStack(Material.REDSTONE_BLOCK)),
@@ -1382,7 +1382,7 @@ public final class BaseItems {
         PylonItem.register(PylonItem.class, SHIMMER_SKULL);
         BasePages.COMPONENTS.addItem(BaseKeys.SHIMMER_SKULL);
 
-        MagicAltar.Recipe.RECIPE_TYPE.addRecipe(new MagicAltar.Recipe(
+        MagicAltarRecipe.RECIPE_TYPE.addRecipe(new MagicAltarRecipe(
                 BaseKeys.SHIMMER_SKULL,
                 new ArrayList<>(Arrays.asList(
                         new RecipeChoice.ExactChoice(SHIMMER_DUST_3),
