+++ src/main/java/io/github/pylonmc/pylon/base/content/machines/simple/MagicAltar.java
@@ -2,10 +2,9 @@ package io.github.pylonmc.pylon.base.content.machines.simple;
 
 import com.destroystokyo.paper.ParticleBuilder;
 import io.github.pylonmc.pylon.base.BaseKeys;
-import io.github.pylonmc.pylon.base.PylonBase;
-import io.github.pylonmc.pylon.base.BaseItems;
 import io.github.pylonmc.pylon.base.content.building.Pedestal;
 import io.github.pylonmc.pylon.base.entities.SimpleItemDisplay;
+import io.github.pylonmc.pylon.base.recipes.MagicAltarRecipe;
 import io.github.pylonmc.pylon.core.block.BlockStorage;
 import io.github.pylonmc.pylon.core.block.PylonBlock;
 import io.github.pylonmc.pylon.core.block.base.PylonInteractableBlock;
@@ -17,15 +16,7 @@ import io.github.pylonmc.pylon.core.entity.display.ItemDisplayBuilder;
 import io.github.pylonmc.pylon.core.entity.display.transform.TransformBuilder;
 import io.github.pylonmc.pylon.core.event.PrePylonCraftEvent;
 import io.github.pylonmc.pylon.core.event.PylonCraftEvent;
-import io.github.pylonmc.pylon.core.guide.button.ItemButton;
-import io.github.pylonmc.pylon.core.i18n.PylonArgument;
-import io.github.pylonmc.pylon.core.item.builder.ItemStackBuilder;
-import io.github.pylonmc.pylon.core.recipe.PylonRecipe;
-import io.github.pylonmc.pylon.core.recipe.RecipeType;
-import io.github.pylonmc.pylon.core.registry.PylonRegistry;
 import io.github.pylonmc.pylon.core.util.PdcUtils;
-import io.github.pylonmc.pylon.core.util.gui.GuiItems;
-import io.github.pylonmc.pylon.core.util.gui.unit.UnitFormat;
 import org.bukkit.*;
 import org.bukkit.block.Block;
 import org.bukkit.entity.ItemDisplay;
@@ -33,13 +24,11 @@ import org.bukkit.event.block.Action;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.RecipeChoice;
 import org.bukkit.persistence.PersistentDataContainer;
 import org.bukkit.util.Vector;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.joml.Vector3i;
-import xyz.xenondevs.invui.gui.Gui;
 
 import java.util.*;
 import java.util.function.Consumer;
@@ -48,7 +37,7 @@ import static io.github.pylonmc.pylon.base.util.BaseUtils.baseKey;
 
 public class MagicAltar extends PylonBlock implements PylonSimpleMultiblock, PylonTickingBlock, PylonInteractableBlock {
 
-    private static final int PEDESTAL_COUNT = 8;
+    public static final int PEDESTAL_COUNT = 8;
 
     private static final NamespacedKey PROCESSING_RECIPE = baseKey("processing_recipe");
     private static final NamespacedKey REMAINING_TIME_SECONDS = baseKey("remaining_time_seconds");
@@ -137,9 +126,9 @@ public class MagicAltar extends PylonBlock implements PylonSimpleMultiblock, Pyl
                 ingredients.add(pedestal.getItemDisplay().getItemStack());
             }
 
-            for (Recipe recipe : Recipe.RECIPE_TYPE.getRecipes()) {
+            for (MagicAltarRecipe recipe : MagicAltarRecipe.RECIPE_TYPE.getRecipes()) {
                 if (recipe.isValidRecipe(ingredients, catalyst)) {
-                    if (!new PrePylonCraftEvent<>(Recipe.RECIPE_TYPE, recipe, this, event.getPlayer()).callEvent()) {
+                    if (!new PrePylonCraftEvent<>(MagicAltarRecipe.RECIPE_TYPE, recipe, this, event.getPlayer()).callEvent()) {
                         continue;
                     }
 
@@ -186,23 +175,23 @@ public class MagicAltar extends PylonBlock implements PylonSimpleMultiblock, Pyl
         return pedestals;
     }
 
-    public @Nullable Recipe getCurrentRecipe() {
+    public @Nullable MagicAltarRecipe getCurrentRecipe() {
         if (processingRecipe == null) {
             return null;
         }
-        return Recipe.RECIPE_TYPE.getRecipe(processingRecipe);
+        return MagicAltarRecipe.RECIPE_TYPE.getRecipe(processingRecipe);
     }
 
     public SimpleItemDisplay getItemDisplay() {
         return getHeldEntityOrThrow(SimpleItemDisplay.class, "item");
     }
 
-    public void startRecipe(Recipe recipe) {
+    public void startRecipe(MagicAltarRecipe recipe) {
         for (Pedestal pedestal : getPedestals()) {
             pedestal.setLocked(true);
         }
-        processingRecipe = recipe.key;
-        remainingTimeSeconds = recipe.timeSeconds;
+        processingRecipe = recipe.key();
+        remainingTimeSeconds = recipe.timeSeconds();
     }
 
     public void tickRecipe(double deltaSeconds) {
@@ -255,9 +244,9 @@ public class MagicAltar extends PylonBlock implements PylonSimpleMultiblock, Pyl
             pedestal.setLocked(false);
         }
 
-        new PylonCraftEvent<>(Recipe.RECIPE_TYPE, getCurrentRecipe(), this).callEvent();
+        new PylonCraftEvent<>(MagicAltarRecipe.RECIPE_TYPE, getCurrentRecipe(), this).callEvent();
 
-        itemDisplay.setItemStack(getCurrentRecipe().result);
+        itemDisplay.setItemStack(getCurrentRecipe().result());
         processingRecipe = null;
 
         new ParticleBuilder(Particle.CAMPFIRE_COSY_SMOKE)
@@ -299,107 +288,6 @@ public class MagicAltar extends PylonBlock implements PylonSimpleMultiblock, Pyl
             spawnParticle.accept(current);
             currentPoint += spacing;
             current.add(step);
-        }
-    }
-
-    /**
-     * Ingredients list must be of size 8. Set an ingredient to null to leave that pedestal empty.
-     *
-     * Ingredients and catalyst must have an amount of 1
-     */
-    public record Recipe(
-            NamespacedKey key,
-            List<RecipeChoice> ingredients,
-            RecipeChoice catalyst,
-            ItemStack result,
-            double timeSeconds
-    ) implements PylonRecipe {
-
-        @Override
-        public @NotNull NamespacedKey getKey() {
-            return key;
-        }
-
-        public static final RecipeType<Recipe> RECIPE_TYPE = new RecipeType<>(
-                new NamespacedKey(PylonBase.getInstance(), "magic_altar")
-        );
-
-        static {
-            PylonRegistry.RECIPE_TYPES.register(RECIPE_TYPE);
-        }
-
-        public boolean ingredientsMatch(List<ItemStack> ingredients) {
-            assert this.ingredients.size() == PEDESTAL_COUNT;
-            assert ingredients.size() == PEDESTAL_COUNT;
-
-            for (int i = 0; i < PEDESTAL_COUNT; i++) {
-
-                boolean allIngredientsMatch = true;
-                for (int j = 0; j < PEDESTAL_COUNT; j++) {
-                    RecipeChoice recipeChoice = this.ingredients.get(j);
-                    if (recipeChoice != null && !recipeChoice.test(ingredients.get(j))) {
-                        allIngredientsMatch = false;
-                        break;
-                    }
-                }
-
-                if (allIngredientsMatch) {
-                    return true;
-                }
-
-                ingredients.add(ingredients.removeFirst());
-            }
-
-            return false;
-        }
-
-        public boolean isValidRecipe(List<ItemStack> ingredients, ItemStack catalyst) {
-            return ingredientsMatch(ingredients) && this.catalyst.test(catalyst);
-        }
-
-        @Override
-        public @NotNull List<@NotNull RecipeChoice> getInputItems() {
-            List<RecipeChoice> choices = new ArrayList<>(ingredients);
-            choices.add(catalyst);
-            choices.removeIf(Objects::isNull);
-            return choices;
-        }
-
-        @Override
-        public @NotNull List<@NotNull ItemStack> getOutputItems() {
-            return List.of(result);
-        }
-
-        @Override
-        public @NotNull Gui display() {
-            return Gui.normal()
-                    .setStructure(
-                            "# # # # # # # # #",
-                            "# # 7 0 1 # # # #",
-                            "m # 6 c 2 # t # r",
-                            "# # 5 4 3 # # # #",
-                            "# # # # # # # # #"
-                    )
-                    .addIngredient('#', GuiItems.backgroundBlack())
-                    .addIngredient('m', ItemButton.fromStack(BaseItems.MAGIC_ALTAR))
-                    .addIngredient('c', ItemButton.fromChoice(catalyst))
-                    .addIngredient('0', ItemButton.fromChoice(ingredients.get(0)))
-                    .addIngredient('1', ItemButton.fromChoice(ingredients.get(1)))
-                    .addIngredient('2', ItemButton.fromChoice(ingredients.get(2)))
-                    .addIngredient('3', ItemButton.fromChoice(ingredients.get(3)))
-                    .addIngredient('4', ItemButton.fromChoice(ingredients.get(4)))
-                    .addIngredient('5', ItemButton.fromChoice(ingredients.get(5)))
-                    .addIngredient('6', ItemButton.fromChoice(ingredients.get(6)))
-                    .addIngredient('7', ItemButton.fromChoice(ingredients.get(7)))
-                    .addIngredient('t', GuiItems.progressCyclingItem((int) (timeSeconds * 20),
-                            ItemStackBuilder.of(Material.CLOCK)
-                                    .name(net.kyori.adventure.text.Component.translatable(
-                                            "pylon.pylonbase.guide.recipe.magic-altar",
-                                            PylonArgument.of("time", UnitFormat.SECONDS.format(timeSeconds))
-                                    ))
-                    ))
-                    .addIngredient('r', ItemButton.fromStack(result))
-                    .build();
         }
     }
 }
