+++ src/main/java/io/github/pylonmc/pylon/base/recipes/MeltingRecipe.java
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.base.content.machines.smelting;
+package io.github.pylonmc.pylon.base.recipes;
 
 import io.github.pylonmc.pylon.core.fluid.PylonFluid;
 import io.github.pylonmc.pylon.core.recipe.PylonRecipe;
@@ -7,37 +7,25 @@ import org.bukkit.NamespacedKey;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.RecipeChoice;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import xyz.xenondevs.invui.gui.Gui;
 
 import java.util.List;
 
 import static io.github.pylonmc.pylon.base.util.BaseUtils.baseKey;
 
-public record CastingRecipe(
+public record MeltingRecipe(
         @NotNull NamespacedKey key,
-        @NotNull PylonFluid fluid,
-        @NotNull ItemStack result,
+        @NotNull ItemStack input,
+        @NotNull PylonFluid result,
         double temperature
 ) implements PylonRecipe {
 
-    public static final double CAST_AMOUNT = 250;
-
-    public static final RecipeType<CastingRecipe> RECIPE_TYPE = new RecipeType<>(
-            baseKey("cast_recipe")
+    public static final RecipeType<MeltingRecipe> RECIPE_TYPE = new RecipeType<>(
+            baseKey("melt_recipe")
     );
 
     static {
         RECIPE_TYPE.register();
-    }
-
-    public static @Nullable CastingRecipe getCastRecipeFor(@NotNull PylonFluid fluid) {
-        for (CastingRecipe recipe : RECIPE_TYPE) {
-            if (recipe.fluid.equals(fluid)) {
-                return recipe;
-            }
-        }
-        return null;
     }
 
     @Override
@@ -47,22 +35,22 @@ public record CastingRecipe(
 
     @Override
     public @NotNull List<@NotNull RecipeChoice> getInputItems() {
-        return List.of();
+        return List.of(new RecipeChoice.ExactChoice(input));
     }
 
     @Override
     public @NotNull List<@NotNull PylonFluid> getInputFluids() {
-        return List.of(fluid);
+        return List.of();
     }
 
     @Override
     public @NotNull List<@NotNull ItemStack> getOutputItems() {
-        return List.of(result);
+        return List.of();
     }
 
     @Override
     public @NotNull List<@NotNull PylonFluid> getOutputFluids() {
-        return List.of();
+        return List.of(result);
     }
 
     @Override
